backport bugfix std::pow

+++ src/EETimingTask.cc
@@ -1,8 +1,8 @@
 /*
  * \file EETimingTask.cc
  *
- * $Date: 2011/08/30 09:28:42 $
- * $Revision: 1.78 $
+ * $Date: 2011/08/30 14:12:37 $
+ * $Revision: 1.78.2.1 $
  * \author G. Della Ricca
  *
 */
@@ -129,7 +129,7 @@ void EETimingTask::setup(void){
   const float maxlogE = 2.;
   float binEdgesE[nbinsE + 1];
   for(int i = 0; i <= nbinsE; i++)
-    binEdgesE[i] = std::pow(10., minlogE + (maxlogE - minlogE) / nbinsE * i);
+    binEdgesE[i] = std::pow((float)10., minlogE + (maxlogE - minlogE) / nbinsE * i);
 
   const int nbinsT = 200;
   const float minT = -50.;
